move error to file
